requirements.txt
mysql-connector-python>=8.0.0
pytest>=7.0.0


